requirements.txt
mysql_connector_repackaged==0.3.1
names==0.3.0
randomname==0.2.1
